Consolidate write result classes into WriteResult
