--- docs/api/MongoDB/Write/InsertResult.php
@@ -1,62 +1,0 @@
-<?php
-
-namespace MongoDB\Write;
-
-/**
- * Result returned by Server and Manager executeWrite() methods processing an
- * InsertBatch.
- *
- * This class may be constructed internally if it will encapsulate a libmongoc
- * data structure.
- */
-final class InsertResult implements WriteResult
-{
-    /**
-     * Returns the Number of documents that where inserted
-     *
-     * @return integer
-     */
-    public function getNumInserted() {}
-
-    /**
-     * Returns all generated IDs
-     *
-     * @return GeneratedId[]
-     */
-    public function getGeneratedIds()
-    {
-        /* Return an array of identifiers generated by the driver. Each
-         * GeneratedId has a batch index and the ID value.
-         *
-         * This supersedes the identifier-setting hack in the 1.x driver.
-         */
-    }
-
-    /**
-     * Returns metadata about the operation, see https://github.com/mongodb/specifications/blob/master/source/server_write_commands.rst#situational-fields
-     *
-     * @return array Additional metadata for the operation(s) (e.g. lastOp)
-     */
-    public function getInfo() {}
-
-    /**
-     * Returns the Server object that this result originated
-     *
-     * @return Server Server from which the result originated
-     */
-    public function getServer() {}
-
-    /**
-     * Returns all WriteConcern Errors that occurred
-     *
-     * @return WriteConcernError[]
-     */
-    public function getWriteConcernErrors() {}
-
-    /**
-     * Returns all Write Errors that occurred
-     *
-     * @return WriteError[]
-     */
-    public function getWriteErrors() {}
-}
