+++ docs/api/MongoDB/Write/WriteResult.php
@@ -2,33 +2,101 @@
 
 namespace MongoDB\Write;
 
-interface WriteResult
+/**
+ * Result returned by Server and Manager executeWrite() methods.
+ *
+ * This class may be constructed internally if it will encapsulate a libmongoc
+ * data structure.
+ */
+final class WriteResult
 {
     /**
-     * Returns metadata about the operation, see https://github.com/mongodb/specifications/blob/master/source/server_write_commands.rst#situational-fields
+     * Returns the GeneratedIds for any inserted documents
      *
+     * @return GeneratedId[]
+     */
+    public function getGeneratedIdsForInsert()
+    {
+        /* Return an array of identifiers generated by the driver for insert
+         * operations. Each GeneratedId has a batch index and the ID value.
+         *
+         * This supersedes the identifier-setting hack in the 1.x driver.
+         */
+    }
+
+    /**
+     * Returns the GeneratedIds for any upserted documents
+     *
+     * @return GeneratedId[]
+     */
+    public function getGeneratedIdsForUpsert()
+    {
+        /* Return an array of identifiers generated by the server for upsert
+         * operations. Each GeneratedId has a batch index and the ID value.
+         */
+    }
+
+    /**
+     * Returns the number of documents that were inserted
+     *
+     * @return integer
+     */
+    public function getNumInserted() {}
+
+    /**
+     * Returns the number of documents that matched the update criteria
+     *
+     * @return integer
+     */
+    public function getNumMatched() {}
+
+    /**
+     * Returns the number of documents that were actually modified by an update
+     *
+     * @return integer
+     */
+    public function getNumModified() {}
+
+    /**
+     * Returns the number of documents that were deleted
+     *
+     * @return integer
+     */
+    public function getNumRemoved() {}
+
+    /**
+     * Returns the number of documents that were upserted
+     *
+     * @return integer
+     */
+    public function getNumUpserted() {}
+
+    /**
+     * Returns metadata about the operation.
+     *
+     * @see https://github.com/mongodb/specifications/blob/master/source/server_write_commands.rst#situational-fields
      * @return array Additional metadata for the operation(s) (e.g. lastOp)
      */
-    function getInfo();
+    public function getInfo() {}
 
     /**
-     * Returns the Server object that this result originated
+     * Returns the Server from which the result originated
      *
-     * @return Server Server from which the result originated
+     * @return Server
      */
-    function getServer();
+    public function getServer() {}
 
     /**
-     * Returns all WriteConcern Errors that occurred
+     * Return any write concern errors that occurred
      *
      * @return WriteConcernError[]
      */
-    function getWriteConcernErrors();
+    public function getWriteConcernErrors() {}
 
     /**
-     * Returns all Write Errors that occurred
+     * Returns any write errors that occurred
      *
      * @return WriteError[]
      */
-    function getWriteErrors();
+    public function getWriteErrors() {}
 }
